Create Assignment 2; Array Lists
If I were to go back and do this assignment again, I would instead code in methods that allow the user to call upon these methods whenever they want to add or remove a person from the student list. This would be more efficient since they can remove a student from the list instead of going through the process of adding someone first.